+++ src/misc/coders/Encoders/video/VP9.js
@@ -1,7 +1,11 @@
 import React from 'react';
 
 import Grid from '@mui/material/Grid';
+import MenuItem from '@mui/material/MenuItem';
 
+import { Trans } from '@lingui/macro';
+
+import Select from '../../../Select';
 import Video from '../../settings/Video';
 
 function init(initialState) {
@@ -9,6 +13,7 @@ function init(initialState) {
 		bitrate: '4096',
 		fps: '25',
 		gop: '2',
+		fps_mode: 'passthrough',
 		...initialState,
 	};
 
@@ -42,6 +47,13 @@ function createMapping(settings) {
 		);
 	}
 
+	if (settings.fps_mode !== 'passthrough') {
+		local.push(
+			'-fps_mode',
+			`${settings.fps_mode}`
+		)
+	}
+
 	const mapping = {
 		global: [],
 		local: local,
@@ -49,6 +61,20 @@ function createMapping(settings) {
 
 	return mapping;
 }
+
+function FpsMode(props) {
+	return (
+		<Select label={<Trans>Frames per second mode</Trans>} value={props.value} onChange={props.onChange}>
+			<MenuItem value="passthrough">Frame is passed (Passthrough)</MenuItem>
+			<MenuItem value="cfr">Constant frame rate (CFR)</MenuItem>
+		</Select>
+	);
+}
+
+FpsMode.defaultProps = {
+	value: '',
+	onChange: function (event) {},
+};
 
 function Coder(props) {
 	const settings = init(props.settings);
@@ -82,11 +108,14 @@ function Coder(props) {
 			<Grid item xs={12}>
 				<Video.Bitrate value={settings.bitrate} onChange={update('bitrate')} allowCustom />
 			</Grid>
-			<Grid item xs={12}>
+			<Grid item xs={12} md={6}>
 				<Video.Framerate value={settings.fps} onChange={update('fps')} allowCustom />
 			</Grid>
-			<Grid item xs={12}>
+			<Grid item xs={12} md={6}>
 				<Video.GOP value={settings.gop} onChange={update('gop')} allowAuto allowCustom />
+			</Grid>
+			<Grid item xs={12}>
+				<FpsMode value={settings.fps_mode} onChange={update('fps_mode')} />
 			</Grid>
 		</Grid>
 	);